refactor(#55): Refactor des feeds et de l'event activity
