+++ divertech/app/src/main/java/Si3/divertech/MultiPagesAdminActivity.java
@@ -14,6 +14,10 @@ import com.google.android.material.textfield.TextInputEditText;
 import java.util.Observable;
 import java.util.Observer;
 
+import Si3.divertech.notifications.NotificationCreator;
+import Si3.divertech.notifications.NotificationList;
+import Si3.divertech.users.User;
+
 public class MultiPagesAdminActivity extends AppCompatActivity implements Observer {
 
     @Override
