+++ divertech/app/src/main/java/Si3/divertech/feed/FeedType.java
@@ -1,4 +1,4 @@
-package Si3.divertech;
+package Si3.divertech.feed;
 
 public enum FeedType {
     NOTIFICATION,
