+++ divertech/app/src/main/java/Si3/divertech/notifications/Notification.java
@@ -1,6 +1,8 @@
-package Si3.divertech;
+package Si3.divertech.notifications;
 
 import androidx.annotation.NonNull;
+
+import Si3.divertech.NotificationTypes;
 
 public class Notification {
     private String id;
