+++ divertech/app/src/main/java/Si3/divertech/qr_reader/CameraPreviewActivity.java
@@ -42,9 +42,9 @@ import com.google.mlkit.common.MlKitException;
 
 import Si3.divertech.DataBaseListener;
 import Si3.divertech.DataBaseResponses;
-import Si3.divertech.EventActivity;
-import Si3.divertech.ListEvent;
 import Si3.divertech.R;
+import Si3.divertech.events.EventActivity;
+import Si3.divertech.events.EventList;
 import Si3.divertech.qr_reader.barcodescanner.BarcodeScannerProcessor;
 import Si3.divertech.qr_reader.preference.PreferenceUtils;
 
@@ -330,8 +330,8 @@ public final class CameraPreviewActivity extends AppCompatActivity implements QR
     @Override
     public void onDataReceived(String eventId) {
         Log.d("RECEIVED_DATA", eventId);
-        if (!ListEvent.getInstance().containsEvent(eventId))
-            ListEvent.getInstance().eventExists(eventId, this);
+        if (!EventList.getInstance().containsEvent(eventId))
+            EventList.getInstance().eventExists(eventId, this);
         else
             goToEventActivity(eventId);
     }
