+++ divertech/app/src/main/java/Si3/divertech/users/UserData.java
@@ -1,4 +1,4 @@
-package Si3.divertech;
+package Si3.divertech.users;
 
 import android.util.Log;
