+++ divertech/app/src/main/java/Si3/divertech/users/UserType.java
@@ -1,0 +1,5 @@
+package Si3.divertech.users;
+
+public enum UserType {
+    ADMIN, NORMAL
+}
