+++ divertech/app/src/main/res/layout/activity_event.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".EventActivity">
+    tools:context=".events.EventActivity">
 
     <!-- Image et flèche retour arrière-->
