+++ divertech/app/src/main/res/layout/activity_multi_pages_admin.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventActivity"
+    tools:context=".events.EventActivity"
     android:background="@color/white">
 
     <fragment
